refactor(build): Move workflow config information into its own file (#1001)
